+++ frontend/src/components/appstore/util.ts
@@ -1,0 +1,4 @@
+export function firstSentence(text?: string): string | undefined {
+  if (!text) return undefined;
+  return text.split(/[.!?]/)[0]?.trim() || undefined;
+}
